# [core] Per-bucket glyph and icon atlases