--- a/cmake/core-files.cmake
+++ b/cmake/core-files.cmake
@@ -41,7 +41,6 @@ set(MBGL_CORE_FILES
 
     # geometry
     src/mbgl/geometry/anchor.hpp
-    src/mbgl/geometry/binpack.hpp
     src/mbgl/geometry/debug_font_data.hpp
     src/mbgl/geometry/feature_index.cpp
     src/mbgl/geometry/feature_index.hpp