--- a/src/mbgl/gl/context.hpp
+++ b/src/mbgl/gl/context.hpp
@@ -205,11 +205,12 @@ public:
     State<value::BindVertexBuffer> vertexBuffer;
     State<value::BindElementBuffer> elementBuffer;
 
+    State<value::PixelStorePack> pixelStorePack;
+    State<value::PixelStoreUnpack> pixelStoreUnpack;
+
 #if not MBGL_USE_GLES2
     State<value::PixelZoom> pixelZoom;
     State<value::RasterPos> rasterPos;
-    State<value::PixelStorePack> pixelStorePack;
-    State<value::PixelStoreUnpack> pixelStoreUnpack;
     State<value::PixelTransferDepth> pixelTransferDepth;
     State<value::PixelTransferStencil> pixelTransferStencil;
 #endif // MBGL_USE_GLES2