--- a/src/mbgl/gl/value.hpp
+++ b/src/mbgl/gl/value.hpp
@@ -232,6 +232,20 @@ struct BindVertexArray {
     static Type Get(const Context&);
 };
 
+struct PixelStorePack {
+    using Type = PixelStorageType;
+    static const constexpr Type Default = { 4 };
+    static void Set(const Type&);
+    static Type Get();
+};
+
+struct PixelStoreUnpack {
+    using Type = PixelStorageType;
+    static const constexpr Type Default = { 4 };
+    static void Set(const Type&);
+    static Type Get();
+};
+
 #if not MBGL_USE_GLES2
 
 struct PointSize {
@@ -270,20 +284,6 @@ struct RasterPos {
 constexpr bool operator!=(const RasterPos::Type& a, const RasterPos::Type& b) {
     return a.x != b.x || a.y != b.y || a.z != b.z || a.w != b.w;
 }
-
-struct PixelStorePack {
-    using Type = PixelStorageType;
-    static const constexpr Type Default = { 4 };
-    static void Set(const Type&);
-    static Type Get();
-};
-
-struct PixelStoreUnpack {
-    using Type = PixelStorageType;
-    static const constexpr Type Default = { 4 };
-    static void Set(const Type&);
-    static Type Get();
-};
 
 struct PixelTransferDepth {
     struct Type {