Rather than having a single glyph atlas texture and a single image atlas texture, use:

* One glyph atlas texture per tile
* One icon atlas texture per tile
* One atlas texture for background/fill/line patterns

This fixes a host of issues. Most prominently, it lifts limits on number and size of images and glyphs. All atlases are allowed to grow as large as they need to be (using shelf-pack-cpp's `autoResize` option). The effective limits are now a function of the maximum texture size (most important factor), unique glyphs/icons used per tile, and unique patterns used. Hopefully the result is high enough not to matter in practice.

Secondly, it allows us to remove the restriction on changing the dimensions of an existing image, which is a prerequisite for full "smart setStyle" support (#7893). The restriction was required because symbol buffer attributes refer to coordinates within the atlas, such that buffer updates and atlas updates must be synchronized. With per-tile atlases, we can do that.

It does increase memory consumption somewhat, as glyphs and icons used in multiple tiles will appear in multiple textures. This increase is offset to some extent by the fact that the previous implementation added all images to the atlas (#9019) and never removed them (#2202).

I started out with per-_bucket_ atlases, which would further increase the effective size limit at the expense of further increasing memory consumption. Per-tile atlases seem more natural and a better balance between segmentation and memory consumption.

This approach diverges from gl-js, which continues to use a single image atlas, and _per-font-stack_ glyph atlases. The latter was implemented in https://github.com/mapbox/mapbox-gl-js/pull/1923 but never ported to native. The approach in this PR has the advantage that it's amenable to making `text-font` data-driven, while gl-js's approach is not.

* Fixes https://github.com/mapbox/mapbox-gl-native/issues/489
* Fixes https://github.com/mapbox/mapbox-gl-native/issues/2202
* Fixes https://github.com/mapbox/mapbox-gl-native/issues/3185
* Fixes https://github.com/mapbox/mapbox-gl-native/issues/3781
* Fixes https://github.com/mapbox/mapbox-gl-native/issues/6536
* Fixes https://github.com/mapbox/mapbox-gl-native/issues/7618
* Fixes https://github.com/mapbox/mapbox-gl-native/issues/9019
* Fixes https://github.com/mapbox/mapbox-gl-native/issues/9122
* Fixes https://github.com/mapbox/mapbox-gl-native/issues/9157

TODO:

* [x] Add only bucket-specific glyphs/icons to atlas
* [x] Or maybe per-tile atlases instead?
* [x] Write integration tests for changing icon size
* [x] Write integration tests for updating pattern
* [x] Get [upstream shelf-pack-cpp PR](https://github.com/mapbox/shelf-pack-cpp/pull/11) merged, released, and integrated
* [x] Quantify performance/memory effects
* [x] Atlas/buckets aren't fully synchronized
* [x] Remove unused binpack implementation