@jfirebaugh 2017-06-09
Memory usage is better than expected, but rendering performance is worse.

Before:

```
* Measured vectorFootprint = 31113762 Bytes for 'Platform=Linux,Compiler=gcc-5,Arch=x86_64'
70.8 MB in fullscreen macosapp
2017-06-09 15:30:21.967988-0700 Bench GL[5781:1398013] | paris      | 53.4 fps |
2017-06-09 15:30:21.968083-0700 Bench GL[5781:1398013] | paris2     | 60.1 fps |
2017-06-09 15:30:21.968167-0700 Bench GL[5781:1398013] | alps       | 32.6 fps |
2017-06-09 15:30:21.968249-0700 Bench GL[5781:1398013] | us east    | 60.0 fps |
2017-06-09 15:30:21.968332-0700 Bench GL[5781:1398013] | greater la | 48.6 fps |
2017-06-09 15:30:21.968606-0700 Bench GL[5781:1398013] | sf         | 60.1 fps |
2017-06-09 15:30:21.968739-0700 Bench GL[5781:1398013] | oakland    | 58.1 fps |
2017-06-09 15:30:21.968826-0700 Bench GL[5781:1398013] | germany    | 31.2 fps |
2017-06-09 15:30:21.968955-0700 Bench GL[5781:1398013] Total FPS: 404.0
2017-06-09 15:30:21.969031-0700 Bench GL[5781:1398013] Average FPS: 50.5
```

After:

```
* Measured vectorFootprint = 18355268 Bytes for 'Platform=Linux,Compiler=gcc-5,Arch=x86_64'
62.6 MB in fullscreen macosapp
2017-06-09 15:39:03.946082-0700 Bench GL[5799:1400783] | paris      | 49.9 fps |
2017-06-09 15:39:03.946237-0700 Bench GL[5799:1400783] | paris2     | 59.4 fps |
2017-06-09 15:39:03.946327-0700 Bench GL[5799:1400783] | alps       | 30.8 fps |
2017-06-09 15:39:03.946587-0700 Bench GL[5799:1400783] | us east    | 58.7 fps |
2017-06-09 15:39:03.946728-0700 Bench GL[5799:1400783] | greater la | 41.2 fps |
2017-06-09 15:39:03.946899-0700 Bench GL[5799:1400783] | sf         | 59.9 fps |
2017-06-09 15:39:03.947110-0700 Bench GL[5799:1400783] | oakland    | 48.9 fps |
2017-06-09 15:39:03.947198-0700 Bench GL[5799:1400783] | germany    | 25.8 fps |
2017-06-09 15:39:03.947272-0700 Bench GL[5799:1400783] Total FPS: 374.6
2017-06-09 15:39:03.947344-0700 Bench GL[5799:1400783] Average FPS: 46.8
```