@jfirebaugh 2017-06-09
Profiling revealed that the textures were being re-created on every frame. 🙈  After fixing that, performance is the same as before:

```
2017-06-09 16:27:07.437403-0700 Bench GL[5828:1410697] | paris      | 53.0 fps |
2017-06-09 16:27:07.437496-0700 Bench GL[5828:1410697] | paris2     | 60.2 fps |
2017-06-09 16:27:07.437579-0700 Bench GL[5828:1410697] | alps       | 34.5 fps |
2017-06-09 16:27:07.437662-0700 Bench GL[5828:1410697] | us east    | 60.0 fps |
2017-06-09 16:27:07.437743-0700 Bench GL[5828:1410697] | greater la | 48.0 fps |
2017-06-09 16:27:07.437877-0700 Bench GL[5828:1410697] | sf         | 60.0 fps |
2017-06-09 16:27:07.437962-0700 Bench GL[5828:1410697] | oakland    | 57.5 fps |
2017-06-09 16:27:07.438043-0700 Bench GL[5828:1410697] | germany    | 30.3 fps |
2017-06-09 16:27:07.438115-0700 Bench GL[5828:1410697] Total FPS: 403.6
2017-06-09 16:27:07.438186-0700 Bench GL[5828:1410697] Average FPS: 50.5
```